Out with Serverless, in with Elastic Beanstalk
In yet another attempt to mitigate the timeout issues that have plagued
this project (see #6), we are now moving to AWS Elastic Beanstalk.

Elastic Beanstalk will keep at least one instance of the code warm at all
times, so we shouldn't run into the scenario where the last instance of
the code has been put to sleep, causing a timeout on first spin-up.